refactor: Landing Page와 Main Page를 구분
